Move menu data into Setup module
