+++ RX671_MCR/inc/setup.h
@@ -1,25 +1,23 @@
-#ifndef GUI_H_
-#define GUI_H_
+#ifndef SETUP_H_
+#define SETUP_H_
 //====================================//
 // インクルード
 //====================================//
 #include "r_smc_entry.h"
-
 #include "ssd1351.h"
 #include "switch.h"
 #include "battery.h"
 #include "bmi088.h"
 #include "linesensor.h"
 #include "encoder.h"
-#include <stdbool.h>
 #include <stdint.h>
-//=====================================//
+#include <stdbool.h>
+
+//====================================//
 // シンボル定義
-//=====================================//
+//====================================//
+
 // メニュー表示に関する定数
-// MENU_START_Y : メニューエリアの開始Y座標
-// MENU_ITEM_HEIGHT : 1行あたりの高さ
-// MAX_VISIBLE_ITEMS : 一度に表示できる行数
 #define MENU_START_Y 12
 #define MENU_ITEM_HEIGHT 12
 #define MAX_VISIBLE_ITEMS (((SSD1351_HEIGHT - MENU_START_Y) / MENU_ITEM_HEIGHT) + 1)
@@ -27,7 +25,9 @@
 //====================================//
 // グローバル変数の宣言
 //====================================//
-extern volatile uint32_t cntGUI;	// GUI用カウンタ
+
+extern volatile uint32_t cntGUI;
+
 //====================================//
 // プロトタイプ宣言
 //====================================//
@@ -41,4 +41,6 @@ void GUI_DrawTestPattern(uint8_t y_start);
 bool GUI_DisplayInverse(void);
 bool GUI_ShowQRcode(void);
 bool GUI_ShowSensors(void);
-#endif // GUI_H_
+void SetupUpdate(void);
+
+#endif // SETUP_H_
